[Refactor] Move routine load consumers into data workflows
