+++ be/src/data_workflows/load/routine_load/data_consumer_group.cpp
@@ -31,14 +31,14 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
-#include "runtime/routine_load/data_consumer_group.h"
+#include "data_workflows/load/routine_load/data_consumer_group.h"
 
 #include "base/utility/defer_op.h"
 #include "compute_env/load/stream_load_context.h"
+#include "data_workflows/load/routine_load/data_consumer.h"
+#include "data_workflows/load/routine_load/kafka_consumer_pipe.h"
 #include "librdkafka/rdkafka.h"
 #include "librdkafka/rdkafkacpp.h"
-#include "runtime/routine_load/data_consumer.h"
-#include "runtime/routine_load/kafka_consumer_pipe.h"
 
 namespace starrocks {
