+++ be/src/data_workflows/load/routine_load/data_consumer_group.h
@@ -36,7 +36,7 @@
 
 #include "base/concurrency/blocking_queue.hpp"
 #include "common/thread/priority_thread_pool.hpp"
-#include "runtime/routine_load/data_consumer.h"
+#include "data_workflows/load/routine_load/data_consumer.h"
 
 namespace starrocks {
