+++ be/src/orchestration/routine_load_task_executor.cpp
@@ -45,14 +45,14 @@
 #include "common/status.h"
 #include "compute_env/load/load_stream_mgr.h"
 #include "compute_env/load/stream_load_context.h"
+#include "data_workflows/load/routine_load/data_consumer_group.h"
+#include "data_workflows/load/routine_load/kafka_consumer_pipe.h"
 #include "data_workflows/load/stream_load/stream_load_executor.h"
 #include "orchestration/stream_load_orchestrator.h"
 #include "rapidjson/document.h"
 #include "rapidjson/stringbuffer.h"
 #include "rapidjson/writer.h"
 #include "runtime/exec_env.h"
-#include "runtime/routine_load/data_consumer_group.h"
-#include "runtime/routine_load/kafka_consumer_pipe.h"
 #include "runtime/runtime_metrics.h"
 
 namespace starrocks::orchestration {
