+++ be/test/data_workflows/load/routine_load/data_consumer_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "runtime/routine_load/data_consumer.h"
+#include "data_workflows/load/routine_load/data_consumer.h"
 
 #include <gtest/gtest.h>
